Fix bug with setting caret to end
End position is determined from the value for the text area. This is
controlled by the parent component. The caret should be set when the
current note changes. But ESLint warns that the value should also
be given as a condition to the effect. But this means that the effect
is invoked when the value for the component is changed, which will
always move the caret to the end. Disabled the warning for now; maybe there's
a way to work around but also, maybe this is not a useful ESLint rule.